tests/end2end/helpers: move 'form' from /helpers/screens to /helpers
