+++ tests/end2end/helpers/screens/document/form_edit_grammar.py
@@ -2,7 +2,7 @@
 from selenium.webdriver.common.by import By
 from seleniumbase import BaseCase
 
-from tests.end2end.helpers.screens.form.form import Form
+from tests.end2end.helpers.form.form import Form
 
 
 class Form_EditGrammar(Form):  # pylint: disable=invalid-name
