+++ tests/end2end/helpers/screens/document/form_edit_requirement.py
@@ -1,7 +1,7 @@
 # pylint: disable=invalid-name
 from seleniumbase import BaseCase
 
-from tests.end2end.helpers.screens.form.form import Form
+from tests.end2end.helpers.form.form import Form
 
 
 class Form_EditRequirement(Form):  # pylint: disable=invalid-name
